Document and test multi-key semantics
Canton relies on the Engine working correctly even in a setting where
we do not have multiple keys. So far this worked by accident but the
semantics of this are rather unclear. To make things worse, Canton
upgrades rely on those semantics being stable so we really do care
about the choices we make here.

This PR does deliberately not change any semantics and I’m not going
to add any semantic changes in here. Instead, it adds a whole suite of
tests for the current behavior which we can use to reach agreement on
whether any of this should change in future PRs.

I’ve also added some documentation on how things should behave in uck
mode both as comments in the tests and as a comment in the
code. However, again as before I’m not going to change any semantics
in this PR so the actual check for that does not yet exist.

changelog_begin
changelog_end